Split the gruntfile into multiple components
