+++ package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "backbone": "~1.1.2",
-    "grunt": "~0.4",
+    "grunt": "~0.4.5",
     "grunt-cli": "~0.1",
     "grunt-contrib-jade": "0.11.0",
     "grunt-contrib-stylus": "~0.5.0",
